# desktop-app/openapi/stocktrader_desktop.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StockTrader | desktop",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/custom-patterns/preview": {
      "post": {
        "tags": [
          "PatternPreviewEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatternPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomPatternWriteRequest": {
        "type": "object",
        "properties": {
          "documentVersion": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "entryRulesJson": {
            "type": "string"
          },
          "entryLogic": {
            "type": "string"
          },
          "requireBullRegime": {
            "type": "boolean"
          },
          "atrStopMultiplier": {
            "type": "number",
            "format": "double"
          },
          "atrTargetMultiplier": {
            "type": "number",
            "format": "double"
          },
          "maxHoldingBars": {
            "type": "integer",
            "format": "int32"
          },
          "trailingAtr": {
            "type": "number",
            "format": "double"
          },
          "partialProfitR": {
            "type": "number",
            "format": "double"
          },
          "useWeightTiers": {
            "type": "boolean"
          },
          "weightTiersJson": {
            "type": "string"
          },
          "defaultAllocationPercent": {
            "type": "number",
            "format": "double"
          },
          "exitRulesJson": {
            "type": "string"
          },
          "exitRulesLogic": {
            "type": "string"
          },
          "exitGroupsJson": {
            "type": "string"
          },
          "exitGroupsLogic": {
            "type": "string"
          },
          "scalingRulesJson": {
            "type": "string"
          },
          "timeFilterJson": {
            "type": "string"
          },
          "circuitBreakerJson": {
            "type": "string"
          },
          "reentryJson": {
            "type": "string"
          },
          "portfolioRulesJson": {
            "type": "string"
          },
          "entryGroupsJson": {
            "type": "string"
          },
          "entryGroupsLogic": {
            "type": "string"
          },
          "dynamicExitJson": {
            "type": "string"
          },
          "entryMode": {
            "type": "string"
          },
          "timeFrame": {
            "$ref": "#/components/schemas/TimeFrame"
          },
          "sizingMode": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "enableLiveTrading": {
            "type": "boolean"
          }
        }
      },
...
      "PatternPreviewRequest": {
        "required": [
          "symbol",
          "pattern"
        ],
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "pattern": {
            "$ref": "#/components/schemas/CustomPatternWriteRequest"
          },
          "timeFrame": {
            "$ref": "#/components/schemas/TimeFrame"
          },
          "from": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "to": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "TimeFrame": {
        "enum": [
          "OneMinute",
          "FiveMinute",
          "FifteenMinute",
          "Daily",
          "Weekly"
        ]
      },
...
    }
  },
...
}